requirements.txt
supabase>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
pydantic>=2.0.0
schedule>=1.2.0
python-dateutil>=2.8.0
tenacity>=8.2.0
pytz>=2023.3
openai>=1.0.0
google-search-results>=2.4.0
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.0.0
